+++ src/main/java/coupon/config/DataSourceConfig.java
@@ -1,4 +1,4 @@
-package coupon.datasource;
+package coupon.config;
 
 import java.util.HashMap;
 import java.util.Map;
