+++ src/main/java/coupon/config/DataSourceContextHolder.java
@@ -1,5 +1,9 @@
-package coupon.datasource;
+package coupon.config;
 
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class DataSourceContextHolder {
 
     private static final ThreadLocal<DataSourceType> dataSourceContext = new ThreadLocal<>();
@@ -10,5 +14,9 @@ public class DataSourceContextHolder {
 
     public static void setDataSource(DataSourceType dataSourceType) {
         dataSourceContext.set(dataSourceType);
+    }
+
+    public static void clearDataSource() {
+        dataSourceContext.remove();
     }
 }
